refactor: 화면에 관련된 controller와 api 컨트롤러를 분리
